+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file - [read more
 
 ### Fixed
 - Error when a subclassed integration returns an object that cannot be cast as a string #423
+- initial redis auto instrumentation integration (https://pecl.php.net/package/redis)
 
 ## [0.21.0]